moving out reusable stuff
