+++ feeder/common/config.py
@@ -1,0 +1,11 @@
+def get_req_header():
+    return {
+        "Accept-Language": "en-US,en;q=0.9",
+        "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
+        "Connection": "keep-alive",
+        "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/64.0.3282.186 Safari/537.36"
+    }
+
+API_URL = "https://keralarescue.in/data"
+
+SOLR_URL = "http://localhost:8983/solr/krescue10/select"
